+++ src/App.scss
@@ -1,11 +1,4 @@
 @import "./Variables.scss";
-@import "./components/home/Home.scss";
-@import "./components/home/Menu.scss";
-@import "./components/Settings/Settings.scss";
-@import "./components/Settings/CategoryForm.scss";
-@import "./components/Profile/Profile.scss";
-@import "./components/Settings/Modal.scss";
-@import "./components/NotFound/NotFound.scss";
 
 /*general*/
 :root {
