+++ src/components/Login/Login.js
@@ -1,0 +1,12 @@
+function Login() {
+    return (
+        <div className="login">
+            <h2> Profile </h2>
+            <div>
+                <button className="log-in"> log in</button>
+            </div>
+        </div>
+    );
+}
+
+export default Login;
