+++ src/components/SignUp/SignUp.js
@@ -1,0 +1,118 @@
+import React, { useState } from "react";
+import axios from "axios";
+
+function SignUp() {
+    const [firstName, setFirstName] = useState("");
+    const [lastName, setLastName] = useState("");
+    const [email, setEmail] = useState("");
+    const [error, setError] = useState({
+        firstName: null,
+        lastName: null,
+        email: null,
+    });
+
+    async function onAddUser() {
+        setError({
+            firstName: null,
+            lastName: null,
+            email: null,
+        });
+        if (firstName.length < 1) {
+            setError((error) => ({
+                ...error,
+                firstName: "please fill in your first name.",
+            }));
+        }
+        if (lastName.length < 1) {
+            setError((error) => ({
+                ...error,
+                lastName: "please fill in your last name.",
+            }));
+        }
+        if (email.length < 1) {
+            setError((error) => ({
+                ...error,
+                email: "please fill in your email.",
+            }));
+        }
+        if (firstName.length < 1 || lastName.length < 1 || email.length < 1) {
+            return;
+        }
+
+        try {
+            await axios.post("http://localhost:4000/api/users", {
+                firstName,
+                lastName,
+                email,
+            });
+        } catch (error) {
+            console.error(error);
+        }
+    }
+
+    return (
+        <div className="sign-up">
+            <h2> Profile </h2>
+            <form>
+                <div className="input-container">
+                    <label> first name </label>
+                    <input
+                        placeholder="Jason"
+                        type="text"
+                        id="firstname"
+                        onChange={(e) => {
+                            e.preventDefault();
+                            setFirstName(e.target.value);
+                        }}
+                    />
+                    {error.firstName && (
+                        <p className="error">{error.firstName}</p>
+                    )}
+                </div>
+                <div className="input-container">
+                    <label> last name </label>
+                    <input
+                        placeholder="Doe"
+                        type="text"
+                        id="lastname"
+                        onChange={(e) => {
+                            e.preventDefault();
+                            setLastName(e.target.value);
+                        }}
+                    />
+                    {error.lastName && (
+                        <p className="error">{error.lastName}</p>
+                    )}
+                </div>
+                <div className="input-container">
+                    <label> e-mail </label>
+                    <input
+                        placeholder="@email.com"
+                        type="text"
+                        id="email"
+                        onChange={(e) => {
+                            e.preventDefault();
+                            setEmail(e.target.value);
+                        }}
+                    />
+                    {error.email && <p className="error">{error.email}</p>}
+                </div>
+
+                <div>
+                    <button
+                        className="primary"
+                        type="submit"
+                        onClick={(e) => {
+                            e.preventDefault();
+                            onAddUser(e);
+                        }}
+                    >
+                        sign up
+                    </button>
+                </div>
+            </form>
+        </div>
+    );
+}
+
+export default SignUp;
