+++ src/components/SignUp/SignUp.scss
@@ -1,10 +1,7 @@
-.profile {
+.sign-up {
     margin: 2rem;
     h2 {
         font-weight: 600;
         font-size: 1rem;
-    }
-    button {
-        margin-bottom: 1rem;
     }
 }
