+++ src/components/index.js
@@ -4,5 +4,7 @@ import NotFound from "./NotFound/NotFound";
 import Settings from "./Settings/Settings";
 import CategoryForm from "./Settings/CategoryForm";
 import Modal from "./Settings/Modal";
+import SignUp from "./SignUp/SignUp";
+import Login from "./Login/Login";
 
-export { Home, Menu, NotFound, Settings, Modal };
+export { Home, Menu, NotFound, Settings, CategoryForm, Modal, SignUp, Login };
