+++ src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 
 import App from "./App";
+import "./styles.scss";
 
 ReactDOM.render(
     <BrowserRouter>
