+++ src/styles.scss
@@ -1,0 +1,12 @@
+@import "./Variables.scss";
+@import "./App.scss";
+
+@import "./components/home/Home.scss";
+@import "./components/home/Menu.scss";
+@import "./components/Login/Login.scss";
+@import "./components/NotFound/NotFound.scss";
+@import "./components/Profile/Profile.scss";
+@import "./components/Settings/CategoryForm.scss";
+@import "./components/Settings/Modal.scss";
+@import "./components/Settings/Settings.scss";
+@import "./components/SignUp/SignUp.scss";
